Crate ipci_runtime

The Robonomics runtime. This can be compiled with #[no_std], ready for Wasm.

Modules

api

constants

A set of constant values used in substrate runtime.

Structs

AdjustmentVariable

ApprovalDeposit

AssetDeposit

Author

BagThresholds

BasicDeposit

BlockHashCount

BondingDuration

Burn

DataDepositPerByte

DealWithFees

DisabledValidatorsThreshold

ElectionProviderBenchmarkConfig

ExistentialDeposit

FieldDeposit

GenesisConfig

ImOnlineUnsignedPriority

MaxAdditionalFields

MaxApprovals

MaxAuthorities

MaxKeys

MaxLocks

MaxNominations

MaxNominatorRewardedPerValidator

MaxPeerDataEncodingSize

MaxPeerInHeartbeats

MaxRegistrars

MaxReserves

MaxScheduledPerBlock

MaxSubAccounts

MaximumMessageSize

MaximumSchedulerWeight

MetadataDepositBase

Key = 32 bytes, Value = 36 bytes (32+1+1+1+1) https://github.com/paritytech/substrate/blob/069917b/frame/assets/src/lib.rs#L257L271

MetadataDepositPerByte

MinerMaxIterations

MinerMaxLength

MinerMaxWeight

MinimumMultiplier

MinimumPeriod

MultiPhaseUnsignedPriority

NoPreimagePostponement

NposSolution16

OffchainRandomBalancing

A source of random balance for NposSolver, which is meant to be run by the OCW election miner.

OffchainRepeat

OffendingValidatorsThreshold

OperationalFeeMultiplier

Origin

The runtime origin type representing the origin of a call.

PalletInfo

Provides an implementation of PalletInfo to provide information about the pallet setup in the runtime.

PreimageBaseDeposit

PreimageByteDeposit

PreimageMaxSize

ProposalBond

ProposalBondMinimum

ReservedDmpWeight

RewardCurve

Runtime

RuntimeApi

RuntimeApiImpl

Implements all runtime apis for the client side.

RuntimeBlockLength

RuntimeBlockWeights

SS58Prefix

SessionKeys

SessionOffset

SessionPeriod

SessionsPerEra

SignedDepositBase

SignedDepositByte

SignedMaxSubmissions

SignedPhase

SignedRewardBase

SlashDeferDuration

SolutionImprovementThreshold

SpendPeriod

StakingBenchmarkingConfig

StakingUnsignedPriority

StringLimit

SubAccountDeposit

TargetBlockFullness

TechnicalMaxMembers

TechnicalMaxProposals

TechnicalMotionDuration

TransactionByteFee

TreasuryPalletId

UncleGenerations

UnsignedPhase

Version

VoterSnapshotPerBlock

WeightToFee

WindowSize

Enums

Call

Event

OriginCaller

StakerStatus

Indicates the initial status of the staker.

Constants

MINER_MAX_ITERATIONS

Maximum number of iterations for balancing that will be executed in the embedded OCW miner of election provider multi phase.

VERSION

Standalone runtime version.

WASM_BINARY

WASM_BINARY_BLOATY

Functions

native_version

The version infromation used to identify this runtime when compiled natively.

wasm_binary_unwrap

Wasm binary unwrapped. If built with BUILD_DUMMY_WASM_BINARY, the function panics.

Type Definitions

Address

The address format for describing accounts.

AllPalletsDeprecated

All pallets included in the runtime as a nested tuple of types.

AllPalletsReversedWithSystemFirst

All pallets included in the runtime as a nested tuple of types in reversed order. With the system pallet first.

AllPalletsWithSystem

All pallets included in the runtime as a nested tuple of types.

AllPalletsWithSystemReversed

All pallets included in the runtime as a nested tuple of types in reversed order.

AllPalletsWithoutSystem

All pallets included in the runtime as a nested tuple of types. Excludes the System pallet.

AllPalletsWithoutSystemReversed

All pallets included in the runtime as a nested tuple of types in reversed order. Excludes the System pallet.

Assets

Aura

AuraConfig

AuraExt

AuraExtConfig

AuraId

An Aura authority identifier using S/R 25519 as its crypto.

Authorship

BagsList

Balances

BalancesConfig

Block

Block type as expected by this runtime.

BlockId

BlockId type as expected by this runtime.

CheckedExtrinsic

Extrinsic type that has already been checked.

Datalog

ElectionProviderMultiPhase

Executive

Executive: handles dispatch to the various modules.

Header

Block header type as expected by this runtime.

Historical

Identity

ImOnline

ImOnlineConfig

ImOnlineId

An i’m online identifier using sr25519 as its crypto.

Offences

ParachainInfo

ParachainInfoConfig

ParachainSystem

Preimage

Scheduler

Session

SessionConfig

SignedExtra

The SignedExtension to the basic transaction logic.

Staking

StakingConfig

Sudo

SudoConfig

System

SystemConfig

TechnicalCommittee

TechnicalCommitteeConfig

TechnicalMembership

TechnicalMembershipConfig

Timestamp

TransactionPayment

Treasury

TreasuryConfig

UncheckedExtrinsic

Unchecked extrinsic type as expected by this runtime.

Utility