Crate ipci_runtime

The Robonomics runtime. This can be compiled with #[no_std], ready for Wasm.

Modules

api

constants

A set of constant values used in substrate runtime.

Structs

AdjustmentVariable

ApprovalDeposit

AssetDeposit

BaseFilter

BasicDeposit

BlockHashCount

BlockReward

Burn

CarbonApprovalDeposit

CarbonAssetAccountDeposit

CarbonAssetDeposit

CarbonMetadataDepositBase

CarbonMetadataDepositPerByte

CarbonStringLimit

CooloffPeriod

DataDepositPerByte

DealWithFees

DepositBase

DepositFactor

EnactmentPeriod

ExistentialDeposit

FastTrackVotingPeriod

FieldDeposit

GenesisConfig

InstantAllowed

LaunchPeriod

MaxAdditionalFields

MaxApprovals

MaxAuthorities

MaxLocks

MaxProposals

MaxRegistrars

MaxReserves

MaxScheduledPerBlock

MaxSignatories

MaxSubAccounts

MaxVotes

MaximumMessageSize

MaximumSchedulerWeight

MetadataDepositBase

Key = 32 bytes, Value = 36 bytes (32+1+1+1+1) https://github.com/paritytech/substrate/blob/069917b/frame/assets/src/lib.rs#L257L271

MetadataDepositPerByte

MinimumDeposit

MinimumMultiplier

MinimumPeriod

NoPreimagePostponement

OperationalFeeMultiplier

Origin

The runtime origin type representing the origin of a call.

PalletInfo

Provides an implementation of PalletInfo to provide information about the pallet setup in the runtime.

PreimageBaseDeposit

PreimageByteDeposit

PreimageMaxSize

ProposalBond

ProposalBondMinimum

ReservedDmpWeight

Runtime

RuntimeApi

RuntimeApiImpl

Implements all runtime apis for the client side.

RuntimeBlockLength

RuntimeBlockWeights

SS58Prefix

SessionKeys

SpendPeriod

StringLimit

SubAccountDeposit

TargetBlockFullness

TechnicalMaxMembers

TechnicalMaxProposals

TechnicalMotionDuration

TransactionByteFee

TreasuryPalletId

Version

VotingPeriod

WeightToFee

WindowSize

Enums

Call

Event

OriginCaller

Constants

VERSION

Standalone runtime version.

WASM_BINARY

WASM_BINARY_BLOATY

Functions

native_version

The version infromation used to identify this runtime when compiled natively.

wasm_binary_unwrap

Wasm binary unwrapped. If built with BUILD_DUMMY_WASM_BINARY, the function panics.

Type Definitions

Address

The address format for describing accounts.

AllPalletsDeprecated

All pallets included in the runtime as a nested tuple of types.

AllPalletsReversedWithSystemFirst

All pallets included in the runtime as a nested tuple of types in reversed order. With the system pallet first.

AllPalletsWithSystem

All pallets included in the runtime as a nested tuple of types.

AllPalletsWithSystemReversed

All pallets included in the runtime as a nested tuple of types in reversed order.

AllPalletsWithoutSystem

All pallets included in the runtime as a nested tuple of types. Excludes the System pallet.

AllPalletsWithoutSystemReversed

All pallets included in the runtime as a nested tuple of types in reversed order. Excludes the System pallet.

Assets

AssetsConfig

Balances

BalancesConfig

Block

Block type as expected by this runtime.

BlockId

BlockId type as expected by this runtime.

CarbonAssets

CarbonAssetsConfig

CheckedExtrinsic

Extrinsic type that has already been checked.

CumulusXcm

Datalog

Democracy

DemocracyConfig

DmpQueue

Executive

Executive: handles dispatch to the various modules.

Header

Block header type as expected by this runtime.

Identity

Lighthouse

Multisig

ParachainInfo

ParachainInfoConfig

ParachainSystem

ParachainSystemConfig

PolkadotXcm

PolkadotXcmConfig

Preimage

RandomnessCollectiveFlip

Scheduler

SignedExtra

The SignedExtension to the basic transaction logic.

Sudo

SudoConfig

System

SystemConfig

TechnicalCommittee

TechnicalCommitteeConfig

TechnicalMembership

TechnicalMembershipConfig

Timestamp

TransactionPayment

TransactionPaymentConfig

Treasury

TreasuryConfig

UncheckedExtrinsic

Unchecked extrinsic type as expected by this runtime.

Utility

XcmpQueue